move pred dataset
